需求信息-API

+++ scape/core/src/main/java/me/scape/ti/dao/RequirementsSecondCategoryDAO.java
@@ -9,6 +9,6 @@ import me.scape.ti.jpa.GenericDAO;
  * @version 1.0.0
  * @since 2015年4月12日 下午11:11:54
  */
-public interface RequirementsSecondCategoryDAO extends GenericDAO<RequirementsSecondCategoryDO, Long> {
+public interface RequirementsSecondCategoryDAO extends GenericDAO<RequirementsSecondCategoryDO, Integer> {
 
 }
