+++ scape/core/src/main/java/me/scape/ti/dao/RequirementsTopCategoryDAO.java
@@ -9,6 +9,6 @@ import me.scape.ti.jpa.GenericDAO;
  * @version 1.0.0
  * @since 2015年4月12日 下午11:14:08
  */
-public interface RequirementsTopCategoryDAO extends GenericDAO<RequirementsTopCategoryDO, Long> {
+public interface RequirementsTopCategoryDAO extends GenericDAO<RequirementsTopCategoryDO, Integer> {
 
 }
